Split functions into binary, string, and url functions
