+++ velox/functions/prestosql/StringFunctions.h
@@ -15,11 +15,6 @@
  */
 #pragma once
 
-#include <folly/hash/Checksum.h>
-#include <cstdint>
-#define XXH_INLINE_ALL
-#include <xxhash.h>
-
 #include "velox/functions/Udf.h"
 #include "velox/functions/lib/string/StringCore.h"
 #include "velox/functions/lib/string/StringImpl.h"
@@ -49,255 +44,6 @@ struct CodePointFunction {
       int32_t& result,
       const arg_type<Varchar>& inputChar) {
     result = stringImpl::charToCodePoint(inputChar);
-  }
-};
-
-/// crc32(varbinary) → bigint
-/// Return an int64_t checksum calculated using the crc32 method in zlib.
-template <typename T>
-struct CRC32Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<int64_t>& result, const arg_type<Varchar>& input) {
-    result = static_cast<int64_t>(folly::crc32_type(
-        reinterpret_cast<const unsigned char*>(input.data()), input.size()));
-  }
-};
-
-/// xxhash64(varbinary) → varbinary
-/// Return an 8-byte binary to hash64 of input (varbinary such as string)
-template <typename T>
-struct XxHash64Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<Varbinary>& result, const arg_type<Varbinary>& input) {
-    // Seed is set to 0.
-    int64_t hash = folly::Endian::swap64(XXH64(input.data(), input.size(), 0));
-    static const auto kLen = sizeof(int64_t);
-
-    // Resizing output and copy
-    result.resize(kLen);
-    std::memcpy(result.data(), &hash, kLen);
-  }
-};
-
-/// md5(varbinary) → varbinary
-template <typename T>
-struct Md5Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  template <typename TTo, typename TFrom>
-  FOLLY_ALWAYS_INLINE void call(TTo& result, const TFrom& input) {
-    stringImpl::md5(result, input);
-  }
-};
-
-/// sha1(varbinary) -> varbinary
-template <typename T>
-struct Sha1Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<Varbinary>& result, const arg_type<Varbinary>& input) {
-    result.resize(20);
-    folly::ssl::OpenSSLHash::sha1(
-        folly::MutableByteRange((uint8_t*)result.data(), result.size()),
-        folly::ByteRange((const uint8_t*)input.data(), input.size()));
-  }
-};
-
-/// sha256(varbinary) -> varbinary
-template <typename T>
-struct Sha256Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  template <typename TTo, typename TFrom>
-  FOLLY_ALWAYS_INLINE void call(TTo& result, const TFrom& input) {
-    stringImpl::sha256(result, input);
-  }
-};
-
-/// sha512(varbinary) -> varbinary
-template <typename T>
-struct Sha512Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  template <typename TTo, typename TFrom>
-  FOLLY_ALWAYS_INLINE void call(TTo& result, const TFrom& input) {
-    stringImpl::sha512(result, input);
-  }
-};
-
-/// spooky_hash_v2_32(varbinary) -> varbinary
-template <typename T>
-struct SpookyHashV232Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<Varbinary>& result, const arg_type<Varbinary>& input) {
-    // Swap bytes with folly::Endian::swap32 similar to the Java implementation,
-    // Velox and SpookyHash only support little-endian platforms.
-    uint32_t hash = folly::Endian::swap32(
-        folly::hash::SpookyHashV2::Hash32(input.data(), input.size(), 0));
-    static const auto kHashLength = sizeof(int32_t);
-    result.resize(kHashLength);
-    std::memcpy(result.data(), &hash, kHashLength);
-  }
-};
-
-/// spooky_hash_v2_64(varbinary) -> varbinary
-template <typename T>
-struct SpookyHashV264Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<Varbinary>& result, const arg_type<Varbinary>& input) {
-    // Swap bytes with folly::Endian::swap64 similar to the Java implementation,
-    // Velox and SpookyHash only support little-endian platforms.
-    uint64_t hash = folly::Endian::swap64(
-        folly::hash::SpookyHashV2::Hash64(input.data(), input.size(), 0));
-    static const auto kHashLength = sizeof(int64_t);
-    result.resize(kHashLength);
-    std::memcpy(result.data(), &hash, kHashLength);
-  }
-};
-
-/// hmac_sha1(varbinary) -> varbinary
-template <typename T>
-struct HmacSha1Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  template <typename TOuput, typename TInput>
-  FOLLY_ALWAYS_INLINE void
-  call(TOuput& result, const TInput& data, const TInput& key) {
-    stringImpl::hmacSha1(result, key, data);
-  }
-};
-
-/// hmac_sha256(varbinary) -> varbinary
-template <typename T>
-struct HmacSha256Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  template <typename TTo, typename TFrom>
-  FOLLY_ALWAYS_INLINE void
-  call(TTo& result, const TFrom& data, const TFrom& key) {
-    stringImpl::hmacSha256(result, key, data);
-  }
-};
-
-/// hmac_sha512(varbinary) -> varbinary
-template <typename T>
-struct HmacSha512Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  template <typename TTo, typename TFrom>
-  FOLLY_ALWAYS_INLINE void
-  call(TTo& result, const TFrom& data, const TFrom& key) {
-    stringImpl::hmacSha512(result, key, data);
-  }
-};
-
-/// hmac_md5(varbinary) -> varbinary
-template <typename T>
-struct HmacMd5Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varbinary>& result,
-      const arg_type<Varbinary>& data,
-      const arg_type<Varbinary>& key) {
-    stringImpl::hmacMd5(result, key, data);
-  }
-};
-
-template <typename T>
-struct ToHexFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varbinary>& result,
-      const arg_type<Varchar>& input) {
-    stringImpl::toHex(result, input);
-  }
-};
-
-template <typename T>
-struct FromHexFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varchar>& result,
-      const arg_type<Varbinary>& input) {
-    stringImpl::fromHex(result, input);
-  }
-};
-
-template <typename T>
-struct ToBase64Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varchar>& result,
-      const arg_type<Varbinary>& input) {
-    stringImpl::toBase64(result, input);
-  }
-};
-
-template <typename T>
-struct FromBase64Function {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varbinary>& result,
-      const arg_type<Varchar>& input) {
-    stringImpl::fromBase64(result, input);
-  }
-};
-
-template <typename T>
-struct FromBase64UrlFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varbinary>& result,
-      const arg_type<Varchar>& input) {
-    stringImpl::fromBase64Url(result, input);
-  }
-};
-
-template <typename T>
-struct ToBase64UrlFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varchar>& result,
-      const arg_type<Varbinary>& input) {
-    stringImpl::toBase64Url(result, input);
-  }
-};
-
-template <typename T>
-struct UrlEncodeFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varchar>& result,
-      const arg_type<Varbinary>& input) {
-    stringImpl::urlEscape(result, input);
-  }
-};
-
-template <typename T>
-struct UrlDecodeFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE void call(
-      out_type<Varchar>& result,
-      const arg_type<Varbinary>& input) {
-    stringImpl::urlUnescape(result, input);
   }
 };
 
@@ -517,57 +263,5 @@ struct StrLPosFunction : public StrPosFunctionBase<T, true> {};
 
 template <typename T>
 struct StrRPosFunction : public StrPosFunctionBase<T, false> {};
-
-template <typename T>
-struct FromBigEndian32 {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<int32_t>& result, const arg_type<Varbinary>& input) {
-    static constexpr auto kTypeLength = sizeof(int32_t);
-    VELOX_USER_CHECK_EQ(input.size(), kTypeLength, "Expected 4-byte input");
-    memcpy(&result, input.data(), kTypeLength);
-    result = folly::Endian::big(result);
-  }
-};
-
-template <typename T>
-struct ToBigEndian32 {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<Varbinary>& result, const arg_type<int32_t>& input) {
-    static constexpr auto kTypeLength = sizeof(int32_t);
-    auto value = folly::Endian::big(input);
-    result.setNoCopy(
-        StringView(reinterpret_cast<const char*>(&value), kTypeLength));
-  }
-};
-
-template <typename T>
-struct FromBigEndian64 {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<int64_t>& result, const arg_type<Varbinary>& input) {
-    static constexpr auto kTypeLength = sizeof(int64_t);
-    VELOX_USER_CHECK_EQ(input.size(), kTypeLength, "Expected 8-byte input");
-    memcpy(&result, input.data(), kTypeLength);
-    result = folly::Endian::big(result);
-  }
-};
-
-template <typename T>
-struct ToBigEndian64 {
-  VELOX_DEFINE_FUNCTION_TYPES(T);
-
-  FOLLY_ALWAYS_INLINE
-  void call(out_type<Varbinary>& result, const arg_type<int64_t>& input) {
-    static constexpr auto kTypeLength = sizeof(int64_t);
-    auto value = folly::Endian::big(input);
-    result.setNoCopy(
-        StringView(reinterpret_cast<const char*>(&value), kTypeLength));
-  }
-};
 
 } // namespace facebook::velox::functions
