+++ velox/functions/prestosql/registration/RegistrationFunctions.h
@@ -39,6 +39,8 @@ void registerURLFunctions(const std::string& prefix = "");
 
 void registerStringFunctions(const std::string& prefix = "");
 
+void registerBinaryFunctions(const std::string& prefix = "");
+
 void registerBitwiseFunctions(const std::string& prefix = "");
 
 void registerAllScalarFunctions(const std::string& prefix = "");
