+++ velox/functions/prestosql/registration/StringFunctionsRegistration.cpp
@@ -18,7 +18,6 @@
 #include "velox/functions/prestosql/RegexpReplace.h"
 #include "velox/functions/prestosql/SplitPart.h"
 #include "velox/functions/prestosql/StringFunctions.h"
-#include "velox/functions/prestosql/types/JsonType.h"
 
 namespace facebook::velox::functions {
 
@@ -58,37 +57,6 @@ void registerSimpleFunctions(const std::string& prefix) {
   registerFunction<RPadFunction, Varchar, Varchar, int64_t, Varchar>(
       {prefix + "rpad"});
 
-  // Register hash functions.
-  registerFunction<CRC32Function, int64_t, Varbinary>({prefix + "crc32"});
-  registerFunction<XxHash64Function, Varbinary, Varbinary>(
-      {prefix + "xxhash64"});
-  registerFunction<Md5Function, Varbinary, Varbinary>({prefix + "md5"});
-  registerFunction<Sha1Function, Varbinary, Varbinary>({prefix + "sha1"});
-  registerFunction<Sha256Function, Varbinary, Varbinary>({prefix + "sha256"});
-  registerFunction<Sha512Function, Varbinary, Varbinary>({prefix + "sha512"});
-  registerFunction<HmacSha1Function, Varbinary, Varbinary, Varbinary>(
-      {prefix + "hmac_sha1"});
-  registerFunction<HmacSha256Function, Varbinary, Varbinary, Varbinary>(
-      {prefix + "hmac_sha256"});
-  registerFunction<HmacSha512Function, Varbinary, Varbinary, Varbinary>(
-      {prefix + "hmac_sha512"});
-  registerFunction<HmacMd5Function, Varbinary, Varbinary, Varbinary>(
-      {prefix + "hmac_md5"});
-  registerFunction<SpookyHashV232Function, Varbinary, Varbinary>(
-      {prefix + "spooky_hash_v2_32"});
-  registerFunction<SpookyHashV264Function, Varbinary, Varbinary>(
-      {prefix + "spooky_hash_v2_64"});
-
-  registerFunction<ToHexFunction, Varchar, Varbinary>({prefix + "to_hex"});
-  registerFunction<FromHexFunction, Varbinary, Varchar>({prefix + "from_hex"});
-  registerFunction<ToBase64Function, Varchar, Varbinary>(
-      {prefix + "to_base64"});
-  registerFunction<FromBase64Function, Varbinary, Varchar>(
-      {prefix + "from_base64"});
-  registerFunction<ToBase64UrlFunction, Varchar, Varbinary>(
-      {prefix + "to_base64url"});
-  registerFunction<FromBase64UrlFunction, Varbinary, Varchar>(
-      {prefix + "from_base64url"});
   exec::registerStatefulVectorFunction(
       prefix + "like", likeSignatures(), makeLike);
 
@@ -98,15 +66,6 @@ void registerSimpleFunctions(const std::string& prefix) {
       {prefix + "regexp_replace"});
   registerFunction<Re2RegexpReplacePresto, Varchar, Varchar, Varchar, Varchar>(
       {prefix + "regexp_replace"});
-
-  registerFunction<FromBigEndian32, int32_t, Varbinary>(
-      {prefix + "from_big_endian_32"});
-  registerFunction<ToBigEndian32, Varbinary, int32_t>(
-      {prefix + "to_big_endian_32"});
-  registerFunction<FromBigEndian64, int64_t, Varbinary>(
-      {prefix + "from_big_endian_64"});
-  registerFunction<ToBigEndian64, Varbinary, int64_t>(
-      {prefix + "to_big_endian_64"});
 }
 } // namespace
